# enhance(open_numbers): parallelise table imports