--- a/etl/steps/open_numbers.py
+++ b/etl/steps/open_numbers.py
@@ -66,28 +66,37 @@ def add_resource(
     resources: List[frictionless.Resource],
 ) -> None:
     print(f"- {short_name}")
-    if len(resources) > 1:
-        df = load_all_resources(repo.cache_dir, resources)
-    else:
-        (resource,) = resources
-        try:
-            df = resource.to_pandas()
+    try:
+        if len(resources) > 1:
+            df = load_and_combine(repo.cache_dir, resources)
+        else:
+            df = load_table(resources[0])
 
-        except FrictionlessException:
-            # see: https://github.com/owid/etl/issues/36
-            print("  ERROR: skipping")
-            return
-
-        # use smaller, more accurate column types that minimise space
-        if "global" in df.columns:
-            df["geo"] = df.pop("global")
+    except FrictionlessException:
+        # see: https://github.com/owid/etl/issues/36
+        print("  ERROR: skipping")
+        return
 
     t = Table(df)
     t.metadata.short_name = short_name
-    ds.add(utils.underscore_table(t))
+
+    # adapt the table name and its column names to our naming convention
+    t = utils.underscore_table(t)
+
+    ds.add(t)
 
 
-def load_all_resources(
+def load_table(resource: frictionless.Resource) -> pd.DataFrame:
+    df = cast(pd.DataFrame, resource.to_pandas())
+
+    # use smaller, more accurate column types that minimise space
+    if "global" in df.columns:
+        df["geo"] = df.pop("global")
+
+    return df
+
+
+def load_and_combine(
     path: Path, resources: List[frictionless.Resource]
 ) -> pd.DataFrame:
     first = True