Several open_numbers datasets are large, and importing them can take a long time when done table by table.

This PR parallelises each open_numbers dataset import, whilst leaving the overall process serial.

## Benchmarks

On a 10-core M1 Max, running `rm -f data/open_numbers; etl open_numbers`

Before: `2352.86s user 130.22s system 99% cpu 41:35.25 total`
After: `4143.78s user 554.04s system 654% cpu 11:57.23 total` (3.5x faster)