@larsyencken 2022-07-27
Good catch! When I bench it now, `concurrent.futures.ProcessPoolExecutor` is waaaay slower than `multiprocessing.Pool`, probably because of batching behaviour.

I'll check the final numbers, and will only merge this if it's substantially faster than the single-threaded version.