@larsyencken 2022-07-27
@Marigold It occurred to me that now that we might run into memory issues on the analytics server now after this change. If we do, we can revert it.